requirements.txt
alembic==1.0.3
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
blinker==1.4
Bootstrap-Flask==1.0.9
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.4.1
dominate==2.3.4
Faker==1.0.1
Flask==1.0.2
Flask-HTTPAuth==3.2.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.3.0
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
more-itertools==5.0.0
pluggy==0.8.1
py==1.7.0
pycparser==2.19
Pygments==2.3.1
PyMySQL==0.9.2
pytest==4.1.1
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
requests==2.21.0
six==1.11.0
SQLAlchemy==1.2.14
text-unidecode==1.2
thulac==0.2.0
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
